Build libvpx using our own BUILD.gn, so we can support Mac Catalyst
The third_party repo contains BUILD.gn files for a number of
third-party projects, including libvpx. We'd rather not fork that for
Signal's RingRTC build as well, but fortunately libvpx is already a
library that can be swapped out with a different build setup. Add a
fixed version of libvpx's BUILD.gn to a new "ringrtc_overrides"
directory until upstream can correctly be built for Catalyst.